+++ src/test/java/domain/player/BettingMoneyTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import java.math.BigDecimal;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -15,33 +16,33 @@ class BettingMoneyTest {
     @ParameterizedTest(name = "배팅 금액은 최소 100원, 최대 1억이다.")
     @ValueSource(ints = {99, 100000001})
     void createBettingMoneyFailTest(int money) {
-        assertThatThrownBy(() -> BettingMoney.from(money))
+        assertThatThrownBy(() -> BettingMoney.from(BigDecimal.valueOf(money)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @ParameterizedTest(name = "배팅 금액은 최소 100원, 최대 1억이다.")
     @ValueSource(ints = {100, 100000000})
     void createBettingMoneySuccessTest(int money) {
-        assertDoesNotThrow(() -> BettingMoney.from(money));
+        assertDoesNotThrow(() -> BettingMoney.from(BigDecimal.valueOf(money)));
     }
 
     @ParameterizedTest(name = "배팅 금액은 100원 단위로 가능하다.")
     @ValueSource(ints = {101, 110010})
     void createBettingMoneyFailByUnitTest(int money) {
-        assertThatThrownBy(() -> BettingMoney.from(money))
+        assertThatThrownBy(() -> BettingMoney.from(BigDecimal.valueOf(money)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @ParameterizedTest(name = "배팅 금액은 100원 단위로 가능하다.")
     @ValueSource(ints = {10000, 110100})
     void createBettingMoneySuccessByUnitTest(int money) {
-        assertDoesNotThrow(() -> BettingMoney.from(money));
+        assertDoesNotThrow(() -> BettingMoney.from(BigDecimal.valueOf(money)));
     }
 
     @ParameterizedTest(name = "배율에 맞게 배팅금액을 돌려받을 수 있다.")
     @MethodSource("moneyAndRatioProvider")
     void receiveBettingMoneySuccessTest(int money, double ratio) {
-        BettingMoney bettingMoney = BettingMoney.from(money);
+        BettingMoney bettingMoney = BettingMoney.from(BigDecimal.valueOf(money));
 
         assertThat(bettingMoney.applyRatio(ratio).getMoney())
                 .isEqualTo((int) (money * ratio));
